requirements.txt
streamlit
google-cloud-storage
google-cloud-firestore
google-cloud-pubsub
tensorflow
keras
keras_preprocessing
numpy

